only admins can delete patients
