+++ src/components/PatientDetails.tsx
@@ -6,6 +6,7 @@ import { getPatient, softDeletePatient } from '../utils/db/patientOperations';
 import type { Patient } from '../types/patient';
 import { VitalSigns } from './VitalSigns';
 import { LoadingSpinner } from './ui/LoadingSpinner';
+import { useRole } from '../contexts/RoleContext';
 import toast from 'react-hot-toast';
 
 export function PatientDetails() {
@@ -14,6 +15,7 @@ export function PatientDetails() {
   const [patient, setPatient] = useState<Patient | null>(null);
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState<string | null>(null);
+  const { role } = useRole();
 
   useEffect(() => {
     async function loadPatient() {
@@ -40,6 +42,11 @@ export function PatientDetails() {
 
   const handleDelete = async () => {
     if (!patient?.id) return;
+    
+    if (role !== 'admin') {
+      toast.error('Only administrators can delete patients');
+      return;
+    }
     
     const confirmed = window.confirm(
       'Are you sure you want to delete this patient record? This action can be undone later.'
@@ -86,13 +93,15 @@ export function PatientDetails() {
           Back to Patient List
         </button>
         <div className="flex items-center space-x-4">
-          <button
-            onClick={handleDelete}
-            className="inline-flex items-center px-3 py-2 border border-red-300 text-sm leading-4 font-medium rounded-md text-red-700 bg-white hover:bg-red-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-red-500"
-          >
-            <Trash2 className="h-4 w-4 mr-2" />
-            Delete Patient
-          </button>
+          {role === 'admin' && (
+            <button
+              onClick={handleDelete}
+              className="inline-flex items-center px-3 py-2 border border-red-300 text-sm leading-4 font-medium rounded-md text-red-700 bg-white hover:bg-red-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-red-500"
+            >
+              <Trash2 className="h-4 w-4 mr-2" />
+              Delete Patient
+            </button>
+          )}
           <span className="text-sm text-gray-500">
             Last updated: {format(new Date(patient.lastUpdated), 'PPpp')}
           </span>
